requirements.txt
a2a-sdk==0.3.25
ag-ui-protocol==0.1.13
agent-framework==1.0.0rc5
agent-framework-a2a==1.0.0b260311
agent-framework-ag-ui==1.0.0b260311
agent-framework-anthropic==1.0.0b260311
agent-framework-azure==0.0.0
agent-framework-azure-ai==1.0.0rc4
agent-framework-azure-ai-search==1.0.0b260311
agent-framework-azurefunctions==1.0.0b260311
agent-framework-bedrock==1.0.0b260311
agent-framework-chatkit==1.0.0b260311
agent-framework-claude==1.0.0b260311
agent-framework-copilotstudio==1.0.0b260311
agent-framework-core==1.0.0rc5
agent-framework-declarative==1.0.0b260311
agent-framework-devui==1.0.0b260311
agent-framework-durabletask==1.0.0b260311
agent-framework-foundry-local==1.0.0b260311
agent-framework-github-copilot==1.0.0b260311
agent-framework-lab==1.0.0b251024
agent-framework-mem0==1.0.0b260311
agent-framework-ollama==1.0.0b260311
agent-framework-orchestrations==1.0.0b260311
agent-framework-purview==1.0.0b260311
agent-framework-redis==1.0.0b260311
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.85.0
anyio==4.12.1
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.4.0
asttokens==3.0.1
async-lru==2.2.0
asyncio==4.0.0
attrs==25.4.0
azure-ai-agents==1.2.0b5
azure-ai-inference==1.0.0b9
azure-ai-projects==2.0.1
azure-common==1.1.28
azure-core==1.38.3
azure-functions==1.25.0b4
azure-functions-durable==1.5.0
azure-identity==1.26.0b2
azure-identity-broker==1.3.0
azure-search-documents==11.7.0b2
azure-storage-blob==12.29.0b1
babel==2.18.0
backoff==2.2.1
beautifulsoup4==4.14.3
bleach==6.3.0
boto3==1.42.69
botocore==1.42.69
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
claude-agent-sdk==0.1.48
click==8.3.1
colorama==0.4.6
comm==0.2.3
cryptography==46.0.5
debugpy==1.8.20
decorator==5.2.1
defusedxml==0.7.1
distro==1.9.0
docstring_parser==0.17.0
durabletask==1.3.0
durabletask.azuremanaged==1.3.0
executing==2.2.1
fastapi==0.135.1
fastjsonschema==2.21.2
foundry-local-sdk==0.5.1
fqdn==1.5.1
frozenlist==1.8.0
furl==2.1.4
github-copilot-sdk==0.1.33rc0
google-api-core==2.30.0
google-auth==2.49.1
googleapis-common-protos==1.73.0
griffe==1.15.0
grpcio==1.78.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
hyperframe==6.1.0
idna==3.11
importlib==1.0.4
importlib_metadata==8.7.1
ipykernel==7.2.0
ipython==9.11.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.8
isodate==0.7.2
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.13.0
jmespath==1.1.0
json5==0.13.0
jsonpath-ng==1.8.0
jsonpointer==3.0.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.8.0
jupyter_core==5.9.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.4
jupyterlab==4.5.6
jupyterlab_pygments==0.3.0
jupyterlab_server==2.28.0
jupyterlab_widgets==3.0.16
lark==1.3.1
MarkupSafe==3.0.3
matplotlib-inline==0.2.1
mcp==1.26.0
mem0ai==1.0.5
microsoft-agents-activity==0.9.0.dev2
microsoft-agents-copilotstudio-client==0.9.0.dev2
microsoft-agents-hosting-core==0.9.0.dev2
mistune==3.2.0
ml_dtypes==0.5.4
MouseInfo==0.1.3
msal==1.35.1
msal-extensions==1.3.1
multidict==6.7.1
nbclient==0.10.4
nbconvert==7.17.0
nbformat==5.10.4
nest-asyncio==1.6.0
notebook==7.5.5
notebook_shim==0.2.4
numpy==2.4.3
ollama==0.6.1
openai==2.28.0
openai-agents==0.12.3
openai-chatkit==1.6.3
opentelemetry-api==1.40.0
opentelemetry-sdk==1.40.0
opentelemetry-semantic-conventions==0.61b0
opentelemetry-semantic-conventions-ai==0.4.13
orderedmultidict==1.0.2
packaging==24.2
pandocfilters==1.5.1
parso==0.8.6
pillow==12.1.1
platformdirs==4.9.4
portalocker==3.2.0
posthog==7.9.12
prometheus_client==0.24.1
prompt_toolkit==3.0.52
propcache==0.4.1
proto-plus==1.27.1
protobuf==5.29.6
psutil==7.2.2
pure_eval==0.2.3
pyasn1==0.6.3
pyasn1_modules==0.4.2
PyAutoGUI==0.9.54
pycparser==3.0.2
PyGetWindow==0.0.9
Pygments==2.19.2
PyJWT==2.12.1
pymsalruntime==0.20.2
PyMsgBox==2.0.1
pyperclip==1.11.0
PyRect==0.2.0
PyScreeze==1.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-json-logger==4.0.0
python-multipart==0.0.22
python-ulid==3.1.0
pytweening==1.2.0
pytz==2026.1.post1
pywin32==311
pywinpty==3.0.3
PyYAML==6.0.3
pyzmq==27.1.0
qdrant-client==1.17.1
redis==7.3.0
redisvl==0.16.0
referencing==0.37.0
requests==2.32.5
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rpds-py==0.30.0
s3transfer==0.16.0
Send2Trash==2.1.0
setuptools==82.0.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.3
SQLAlchemy==2.1.0b1
sse-starlette==3.3.2
stack-data==0.6.3
starlette==1.0.0rc1
tenacity==9.1.4
terminado==0.18.1
tinycss2==1.4.0
tornado==6.5.5
tqdm==4.67.3
traitlets==5.14.3
types-requests==2.32.4.20260107
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uri-template==1.3.0
urllib3==2.6.3
uvicorn==0.42.0
watchfiles==1.1.1
wcwidth==0.6.0
webcolors==25.10.0
webencodings==0.5.1
websocket-client==1.9.0
websockets==16.0
Werkzeug==3.1.6
widgetsnbextension==4.0.15
yarl==1.23.0
zipp==3.23.0
